Allow replacing "/vale" with "/valueFrom"
While we do not usually allow any remove operations in a diff, we have
to support the case where some environment variable is switched from an
explicit value to retrieving the value from an operator via `valueFrom`.

We keep a list of skipped operations and then check if there's a remove
operation for a value in the skipped ones and if we have a matching add
operation to the same path for a valueFrom.