Allow vSphere clusters with leading zero
Most of our ESXi hosts are named with a leading zero in the BB name,
e.g. bb091. In contrast, our vSphere clusters are called
"productionbb91". To get more consistency and ease the generation of
cluster names by ansible automation scripts, we will move the cluster
names to contain a leading zero, e.g. "productionbb091".

Since, as the commit history tells, we cannot simply walk into mordor,
i.e. change the cluster-finding regex, as this would spawn a new
nova-compute node with a different name, which would not be able to
manage the instances assigned to the old nova-compute node, we ignore
the leading zero for the name of the deployment. The cluster-name is
passed as is to the templates anyways.