Support diferent ways to specify scope and jinja2_options
With https://github.com/kubernetes/kubernetes/pull/64267 kubernetes
stopped allowing custom keys to be set in `metadata`, which is what we
used for `scope` and `jinja2_options`. Therefore, we now introduce a key
`options` on the same level as `template`, where both `scope` and
`jinja2_options` can be set. If it's set there, we will use it in favor
of the ones in `metadata` and only fall back to reading from `metadata`
if `options` is not set. This should ensure backwards compatibility with
older templates on older kubernetes clusters.