Reconnect to vcenters
Previously, we would only connect to a vcenter once: when we newly found
it in DNS, e.g. on startup of the vcenter-operator. It could happen,
that this session would break, because vCenter would get rebootet. We
did not try to reconnect ever, but just seemd to ignore that vCenter.

This commit adds a reconnect functionality. We check each connection
before we try to use it and reconnect, if we could not run a simple RPC
call.

This changes the `atexit` functionality, too. We now can't register a
`Disconnect` for every connection, because connections will get
replaced. Instead, we now have to walk over all our connections and
close them.