Don't forget a vcenter if first connect fails
When connecting to a vCenter found via DNS, we only add it to the list
of known vCenters, if we can connect to it. If not, we forget about it
forever. This can be bad if a vCenter is down on startup of the operator
but becomes available later, because we just ignore it until someone
notices.