Don't error out on failed VC connect
Instead of failing the whole operator because one VC is down, we now
catch any Exception raised in the re-connection process and at least try
to handle all vCenters we can currently connect to.

Otherwise, with the last patch, it would be impossible to handle the
other vCenters as we don't just forget about a failed one.