Keep order of rendered templates when applying
We sort the templates by path on returning them from the loaders to the
Configurator. This order was lost again, when the Configurator added the
rendered versions of the templates to the DeploymentState, as we used
dictionaries there. This could mean, that a configmap was only updated
after the deployment, leading to the restarted deployment loading an
outdated configmap.

We now use OrderedDict to preserve the order in which rendered templates
are added to the DeploymentState. According to the docs, OrderedDict can
be used as drop-in replacement for dicts.