Sort updated objects by Kind
We're still facing the problem, that containers restart because of
updated Deployments, but still contain the old config files. We were
able to observe a deployment getting updated before the configmap in one
of our labs, but couldn't reproduce this.

Trying to fix the problem, we now sort the updated objects by Kind, so
we update the ConfigMaps before the Deployments. This should help for
the Deployments' containers to read the updated ConfigMap, when they
restart.

In a previous attempt, we introduced OrderedDict objects to keep the
order, but we're using set operations which can screw with the order
again, as we could validate.