Fix compatibility with kubernetes api 1.21
Kubernetes 1.21 does not support the export argument to various "get" calls.
It was added with the intention of removing values added by the server,
but never worked as intended. Instead the values where filtered manually.

Additional values to the filter list have been added, and
the python library has been upgraded to match that version.