Add ephemeral DS hagroup detection
When a cluster has datastores for the hagroups a and b, we set a setting
on the cluster that defines the regex to be used to find those hagroups.
This will enable a custom feature in Nova, that allows to place the
ephemeral root disk of servers of the same server-group onto different
ephemeral datastores by hagroup.