Poll Nova cells from k8s secret
Since the os-cells API endpoint in Nova was removed with the cell-v1
removal, we cannot poll the cells from there anymore. Instead, Nova will
be deployed with secrets for each cell labeled system=openstack,
component=nova and type=nova-cell containing the information about those
cells. This commit makes the vcenter-operator read these secrets.

The expected content of the secrets is as follows:
  * A key "name" identifying the cell, base64 encoded
  * A key "database_connection" with the DB URL for this cell, base64
    encoded
  * A key "transport_url" with the rabbitmq URL for this cell, base64 encoded

Since we might reconfigure compute hosts with the wrong cell if we could
not retrieve cell information from Nova/k8s, we now discontinue our
current configuration run if an error occurred during reading the cells.
We try not to crash so we can continue our work once the problem got
resolved.