Replace internal get/diff/apply-diff logic with server-side-apply
The diff was not handling the various fields terribly well,
that get added by either new releases of k8s or possible extensions
installed (owner-info, etc...)

The very reason server side apply was invented:
See https://kubernetes.io/docs/reference/using-api/server-side-apply/

Also the logic was buggy, and not properly removing some fields.

With server-side apply the whole logic is move to the server,
which takes care of only updating the resource, if something has
changed.

The main downside being, that the logs do not show the delta
anymore, as the client is not involved it.