Switch to pyproject/ruff
Pyproject.toml: PEP 621
So we do not have a bunch of config-files flying around.

Ruff is quite a bit faster and comprehensive than flake

Generate a pbr.json to keep our changelog logic happy.

pbr isn't yet 100% compatible with PEP 621 and doesn't
seem to generate the pbr.json anymore by default.

While we are at it, add a `--auto-remove` to the purge,
so we also remove the automatically installed dependencies