Use buildkit features for caching
Instead of deleting the caches, we use the new dockerfile syntax
to make docker aware of the cache paths.

For local builds, it should work out of the box, but in the ci
it needs to be adjusted. Possibly it won't make things faster there,
as the build there is distributed