Resolve OpenStack session
We no longer fetch the cell information via Nova's API, because that
endpoint is long gone - but we still set up the session for doing so.

Since we no longer need an OpenStack session, we resolve the code
creating it and thus reduce the needed for credentials for the operator.