Try to replace resource if server-side-apply fails
We have rather small changes, like adding and removing volumes,
which the k8s api cannot cleanly apply.
The replace api call seems to do the proper thing, so we fall back
to that.
Presumably, that would break a rolling release upgrade, but I'd say
that is better than not updating at all and silently running
with the old release.

Also print first what we about to do, so we have that information
in case something fails later.