Support version numbers > 9999
We hard-coded the versions to be of length 4, because we fill it up to 4
chars with zeroes, but at some far point in the future the version might
be higher than 9999. Therefore, we switch to using `removeprefix()` for
removing the service_username_template from the username after having
checked that the username starts with service_username_template.